+++ tests/test_bit_vec.rs
@@ -12,7 +12,7 @@ use rand::rngs::SmallRng;
 use rand::seq::SliceRandom;
 use std::sync::atomic::AtomicUsize;
 use sux::prelude::*;
-use sux::traits::bit_vec::*;
+use sux::traits::bit_vec_ops::*;
 
 #[test]
 fn test() {
@@ -403,7 +403,7 @@ fn test_iter_zeros_one() {
 
 #[test]
 fn test_atomic_iter() {
-    let mut c = AtomicBitVec::new(100);
+    let c = AtomicBitVec::new(100);
     for i in 0..100 {
         c.set(i, i % 2 == 0, Ordering::Relaxed);
     }