+++ tests/test_select_zero_adapt.rs
@@ -9,7 +9,7 @@ use rand::Rng;
 use rand::SeedableRng;
 use rand::rngs::SmallRng;
 use sux::prelude::*;
-use sux::traits::bit_vec::*;
+use sux::traits::bit_vec_ops::*;
 
 #[test]
 fn test() {